Moved binder health indicator to autoconfiguration
Also updated copyright
